# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Agile Metrics Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects/{projectId}/kanban/{period}:
    get:
      operationId: getProjectKanbanMetrics
      tags: [metrics]
      summary: Get weekly Kanban metrics (throughput, cycle time) for non-Scrum projects
      description: >
        Groups completed issues by ISO week and computes throughput, cycle time,
        reopened count, and issue type breakdown per week. Only available for
        Kanban/simple board projects.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProjectId"
        - $ref: "#/components/parameters/PeriodPath"
      responses:
        "200":
          description: Kanban weekly metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KanbanMetricsResponse"
        "400":
          description: Project is a Scrum project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Project not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
    PeriodPath:
      name: period
      in: path
      required: true
      schema:
        type: string
        enum: ["1m", "3m", "6m"]
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    IssueTypeBreakdown:
      type: object
      required: [Story, Bug, Task, Epic, Other]
      properties:
        Story:
          type: integer
        Bug:
          type: integer
        Task:
          type: integer
        Epic:
          type: integer
        Other:
          type: integer
...
    KanbanMetricsResponse:
      type: object
      required: [weeks, summary]
      properties:
        weeks:
          type: array
          items:
            $ref: "#/components/schemas/WeekMetric"
        summary:
          $ref: "#/components/schemas/KanbanMetricsSummary"
    WeekMetric:
      type: object
      required:
        [weekStart, weekLabel, totalIssues, totalStoryPoints, avgCycleTimeDays,
         reopenedCount, breakdown]
      properties:
        weekStart:
          type: string
          format: date
          description: ISO date of the Monday of this week
        weekLabel:
          type: string
          description: Human-readable label like "Mar 2 - Mar 8"
        totalIssues:
          type: integer
          description: Issues completed this week (throughput)
        totalStoryPoints:
          type: number
        avgCycleTimeDays:
          type: ["number", "null"]
        reopenedCount:
          type: integer
        breakdown:
          $ref: "#/components/schemas/IssueTypeBreakdown"
    KanbanMetricsSummary:
      type: object
      required:
        [totalWeeks, avgThroughput, avgCycleTimeDays, totalCompletedIssues, totalCompletedStoryPoints]
      properties:
        totalWeeks:
          type: integer
        avgThroughput:
          type: number
        avgCycleTimeDays:
          type: ["number", "null"]
        totalCompletedIssues:
          type: integer
        totalCompletedStoryPoints:
          type: number